Refine retrier
