+++ internal/cmd/run/cypress.go
@@ -201,10 +201,10 @@ func runCypressInSauce(p cypress.Project, regio region.Region) (int, error) {
 			FailFast:               gFlags.failFast,
 			MetadataSearchStrategy: framework.NewSearchStrategy(p.GetVersion(), p.GetRootDir()),
 			NPMDependencies:        p.GetNpm().Dependencies,
-			Retrier: &retry.BasicRetrier{
+			Retrier: &retry.SauceReportRetrier{
 				VDCReader:       &restoClient,
 				ProjectUploader: &appsClient,
-				CypressProject:  p,
+				Project:         p,
 			},
 		},
 	}
