+++ internal/cmd/run/playwright.go
@@ -216,10 +216,10 @@ func runPlaywrightInSauce(p playwright.Project, regio region.Region) (int, error
 			FailFast:               gFlags.failFast,
 			MetadataSearchStrategy: framework.NewSearchStrategy(p.Playwright.Version, p.RootDir),
 			NPMDependencies:        p.Npm.Dependencies,
-			Retrier: &retry.BasicRetrier{
-				VDCReader:         &restoClient,
-				ProjectUploader:   &appsClient,
-				PlaywrightProject: p,
+			Retrier: &retry.SauceReportRetrier{
+				VDCReader:       &restoClient,
+				ProjectUploader: &appsClient,
+				Project:         &p,
 			},
 		},
 	}
