+++ internal/cucumber/config.go
@@ -13,6 +13,7 @@ import (
 	"github.com/saucelabs/saucectl/internal/insights"
 	"github.com/saucelabs/saucectl/internal/msg"
 	"github.com/saucelabs/saucectl/internal/region"
+	"github.com/saucelabs/saucectl/internal/saucereport"
 )
 
 // Config descriptors.
@@ -320,4 +321,18 @@ func SortByHistory(suites []Suite, history insights.JobHistory) []Suite {
 		}
 	}
 	return res
+}
+
+func (p *Project) FilterFailedTests(suiteIndex int, report saucereport.SauceReport) error {
+	if suiteIndex < 0 || suiteIndex > len(p.Suites) {
+		return errors.New("invalid suite index")
+	}
+	specs := saucereport.GetFailedSpecFiles(report)
+	// if no failed specs found, just keep the original settings
+	if len(specs) == 0 {
+		return nil
+	}
+	p.Suites[suiteIndex].Options.Paths = specs
+
+	return nil
 }
