+++ internal/cypress/v1/config_test.go
@@ -288,50 +288,3 @@ func TestCyress_CleanPackages(t *testing.T) {
 		})
 	}
 }
-
-func TestCypress_SetTestGrep(t *testing.T) {
-	testCases := []struct {
-		name       string
-		project    Project
-		suiteIndex int
-		tests      []string
-		expResult  map[string]string
-	}{
-		{
-			name: "set TestGrep for the first suite in project",
-			project: Project{
-				Suites: []Suite{
-					{
-						Config: SuiteConfig{
-							Env: map[string]string{},
-						},
-					},
-				},
-			},
-			suiteIndex: 0,
-			tests:      []string{"failed test1", "failed test2"},
-			expResult: map[string]string{
-				"grep": "failed test1;failed test2",
-			},
-		},
-		{
-			name: "set TestGrep for the first suite in project w/ empty Env",
-			project: Project{
-				Suites: []Suite{
-					{},
-				},
-			},
-			suiteIndex: 0,
-			tests:      []string{"failed test1", "failed test2"},
-			expResult: map[string]string{
-				"grep": "failed test1;failed test2",
-			},
-		},
-	}
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			tc.project.SetTestGrep(tc.suiteIndex, tc.tests)
-			assert.Equal(t, tc.expResult, tc.project.Suites[tc.suiteIndex].Config.Env)
-		})
-	}
-}
