+++ internal/saucecloud/retry/projector.go
@@ -1,0 +1,7 @@
+package retry
+
+import "github.com/saucelabs/saucectl/internal/saucereport"
+
+type Project interface {
+	FilterFailedTests(suiteIndex int, report saucereport.SauceReport) error
+}
